cmake: Use GNUInstallDirs to provide configurable & correct install dirs
Use the GNUInstallDirs CMake module to provide a standard set of
variables to control install paths. Otherwise, the package is
incorrectly installed into 'lib' (instead of 'lib64') on 64-bit systems.
While this could be solved via providing custom variables for each
directory, it is simpler to reuse what's already in CMake.