Add a new option for requiring BMI2 intrinsics
If in an envioronment where the desired output code will be run on a
broad set of x86 CPUs, BMI2 isn't guaranteed to be available. When
compiling against MSVC, `bzhi` will still be inserted into the
instruction stream because the `SNAPPY_HAVE_BMI2` check will succeeed
(it compiles without issue), but cannot be run by Sandy Bridge and early
CPUs.